[add] remove unused code + rename files
