+++ packages/core/server/content-api/routes/core.ts
@@ -7,12 +7,6 @@ export default [
     handler: 'core.router',
     config: {
       policies: [],
-      find: {
-        middlewares: ['api::page.default-populate'],
-      },
-      findOne: {
-        middlewares: ['api::page.default-populate'],
-      },
     },
   },
 ];
